Fix bug with parsing a local chainfile.
This required replacing the 'for' loop in LiftOverChainFile._load_chains()
with a 'while True' to avoid conflicts with any iteration on the file object
that happens in any methods/constructors delegated to (e.g LiftOverChain() in
this case).

Also fixed some doctests which were failing on my 64-bit Linux machine
because the return types were not long as in the Expected block. Is
there a better way to fix this?